+++ index.html
@@ -295,13 +295,17 @@
 </head>
 
 <body>
+
+    
     <script src="https://cdnjs.cloudflare.com/ajax/libs/axios/1.4.0/axios.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
         inteGrity="sha384-geWF76RCwLtnZ8qwWowPQNguL3RmwHVBC9FhGdlKrxdiJJigb/j/68SIy3Te4Bkz"
         crossorigin="anonymous"></script>
+
+        
     <script src="/assets/scripts/app2.js" type="module"></script>
     <script src="/assets/scripts/postRequests.js" type="module"></script>
-
+    <script src="/assets/scripts/script.js" type="module"></script>
 
     <nav class="navbar bg-dark border-bottom border-bottom-dark" data-bs-theme="dark">
         <div class="container-fluid">
@@ -800,7 +804,7 @@
                 <option value="Portocaliu">Portocaliu</option>
                 <option value="Gri">Gri</option>
             </select>
-            <ul id="recomandari13" class="recomandari"></ul>
+            <ul id="recomandari6" class="recomandari"></ul>
 
             <div class="match" id="match11">
                 <div class="match-info">
@@ -887,7 +891,7 @@
             </table>
         </div>
 
-        <div class="finala" id="etapa7">
+        <div class="finale" id="etapa7">
             <input type="text" id="searchInput7" placeholder="Introduceți o valoare">
             <button onclick="addValueToList(7)">Introdu</button>
             <ul id="recomandari7" class="recomandari"></ul>
@@ -898,7 +902,7 @@
                 <option value="Portocaliu">Portocaliu</option>
                 <option value="Gri">Gri</option>
             </select>
-            <ul id="recomandari14" class="recomandari"></ul>
+            <ul id="recomandari7" class="recomandari"></ul>
 
             <h3>Finala Mica</h3>
             <div class="finala" id="finala-mica">
@@ -982,7 +986,6 @@
 
 
 
-    <script src="/assets/scripts/script.js"></script>
 </body>
